refactor(channels): extract emojiFor to shared format.go
